% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gme_variations2.R
\name{gme_actions_table}
\alias{gme_actions_table}
\title{Mythic Meaning Tables: Actions}
\usage{
gme_actions_table()
}
\description{
Use it to generate two word actions, such as Agree Magic or Fight Adversities.
}
\seealso{
\code{\link{gme_descriptions_table}},
\code{\link{gme_actions_table2}} and
\code{\link{rand_gma}}
}
