% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gme_variations2.R
\name{gme_actions_table2}
\alias{gme_actions_table2}
\title{Mythic Meaning Tables: Extended Actions}
\usage{
gme_actions_table2()
}
\description{
A homebrewed modification of the original \link[=gme_actions_table]{Actions Table}
adding the first word from the \link[=gme_descriptions_table]{Descriptions Table}
in the middle, to generate results such as Persecute Damaged Fears and
Abuse Graceful Bureaucracy.
}
\seealso{
\code{\link{gme_descriptions_table}},
\code{\link{gme_actions_table2}} and
\code{\link{rand_gma}}
}
