% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gme_variations2.R
\name{gme_description_table}
\alias{gme_description_table}
\title{Mythic Meaning Tables: Descriptions}
\usage{
gme_description_table()
}
\description{
Use it to generate two word descriptions, such as Kindly Delightful or
Delightfully Bizarre.
}
\seealso{
\code{\link{gme_actions_table}},
\code{\link{gme_actions_table2}} and
\code{\link{rand_gma}}
}
