% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gme_variations2.R
\name{gme_detail_check}
\alias{gme_detail_check}
\title{Mythic Variations 2 Detail Check}
\usage{
gme_detail_check(cf = 4, debug = F, themes = F)
}
\arguments{
\item{cf}{A number between 3 and 6. Defaults to 4. Mythic Chaos Factor.
Higher values correspond to situations in which your characters are
disfavoured and viceversa.}

\item{debug}{A logical. If \code{TRUE}, the output is a list with the die roll
result included. If \code{FALSE}, only the result is shown. Defaults to \code{FALSE}.}

\item{themes}{A logical. If \code{TRUE}, results related to NPCs are changed
to results related to Themes. Defaults to \code{FALSE}.}
}
\description{
Use it to answer questions that need more
detail than just a yes or no, such as "How does the tavern look?" or
"What does the cache contain?".
}
