% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gme_variations2.R
\name{gme_event_check}
\alias{gme_event_check}
\title{Mythic Event Check}
\usage{
gme_event_check(theme = "standard")
}
\arguments{
\item{theme}{A character. One of "standard" (default), "horror", "mystery" or
"chekhov".}
}
\description{
Including themes from Variations I. Use it when
the \code{\link{Fate Check}} generates a Random Event, when your scene is
Interrupted or whenever you want to throw some randomness.
}
\seealso{
\code{\link{gme_plot_point()}} and \code{\link{gme_turning_point()}}
}
