% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gme_variations2.R
\name{gme_fate_check}
\alias{gme_fate_check}
\title{Mythic Variations 2 Fate Check}
\usage{
gme_fate_check(chaosFactor = 4, chaosAdjust = 2, prob = 0)
}
\arguments{
\item{chaosFactor}{A number between 3 and 6. Defaults to 4. Higher values
correspond to situations in which your characters are disfavoured and
viceversa. Use \code{chaosAdjust} to control it if your Chaos Factor is 3 or 6.}

\item{chaosAdjust}{A number, one of -2 or +2, modifying the likelihood of a
yes in a positive (CF 3) or negative (CF 6) way for the character/s.}

\item{prob}{A number: 0 (50/50 or unsure), +2/-2 (likely / unlikely),
+4/-4 (very likely / very unlikely), +6/-6 (sure thing / no way) or
 +8/-8 (has to be / impossible).}
}
\description{
Use it to generate yes / no answers, with
the possibility of generating exceptional results and random events.
Please consult Mythic Variations II to understand exactly how these results
 are generated.
}
