% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ironsworn.R
\name{iron_dice}
\alias{iron_dice}
\title{Ironsworn Dice Roll}
\usage{
iron_dice(bonus = 0)
}
\arguments{
\item{bonus}{A number to sum to the 1d6.}
}
\value{
Strong hit, weak hit or miss.
}
\description{
Rolls 1d6 and compares it against 2d10.
}
