% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une_conv.R
\name{une_conv}
\alias{une_conv}
\title{UNE Conversation}
\usage{
une_conv(relationship = "neutral", demeanor = sample(1:8, 1))
}
\arguments{
\item{relationship}{The relationship between the PC and the NPC. One of "loved", "friendly", "peaceful", "neutral" (default), "distrustful", "hostile", "hated".}

\item{demeanor}{One of "scheming", "insane", "friendly", "hostile", "inquisitive", "mysterious", "prejudiced" or a number between 1 and 8. Random by default.}
}
\value{
The mood, bearing and focus of the conversation.
}
\description{
Generate the complete conversation of an NPC using the Universal NPC Emulator.
}
