% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une_conv.R
\name{une_conv_bearing}
\alias{une_conv_bearing}
\title{UNE Conversation Bearing}
\usage{
une_conv_bearing(demeanor = sample(1:8, 1))
}
\arguments{
\item{demeanor}{One of "scheming", "insane", "friendly", "hostile", "inquisitive", "mysterious", "prejudiced" or a number between 1 and 8.}
}
\value{
A word signaling the bearing of the NPC.
}
\description{
Generate the bearing in the conversation of an NPC using the Universal NPC Emulator.
}
