% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une_conv.R
\name{une_conv_focus}
\alias{une_conv_focus}
\title{UNE Conversation Focus}
\usage{
une_conv_focus()
}
\value{
A random word with the focus.
}
\description{
Generate the focus of the conversation of an NPC using the Universal NPC Emulator.
}
