% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une_conv.R
\name{une_conv_mood}
\alias{une_conv_mood}
\title{UNE Conversation Mood}
\usage{
une_conv_mood(relationship = "neutral")
}
\arguments{
\item{relationship}{The relationship between the PC and the NPC. One of "loved", "friendly", "peaceful", "neutral" (default), "distrustful", "hostile", "hated".}
}
\value{
The mood of the NPC, chosen between "withdrawn", "guarded", "cautious", "neutral", "sociable", "helpful" or "forthcoming".
}
\description{
Generate conversation mood from an NPC using the Universal NPC Emulator.
}
